pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pic_prompt"
version = "1.2.0"
description = "A library for building imageprompts for various LLM providers"
authors = [
    { name = "Paul Baranowski", email = "paulbaranowski@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "boto3>=1.26.0",
    "pytest-env>=1.1.5",
    "pillow>=10.4.0",
    "litellm>=1.64.1",
]

[project.urls]
"Homepage" = "https://github.com/paulbaranowski/prompt-any"
"Bug Tracker" = "https://github.com/paulbaranowski/prompt-any/issues"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.20.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/pic_prompt"]

[dependency-groups]
dev = [
    "black>=24.8.0",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "pass",
    "abc.abstractmethod",
    "@abstractmethod"
]
